New changes plus buttons
So I added a new loop for the pre-menu of your game and included a button class to work with. I tried to add some comments but hopefully you can figure this out and implement it on your own. I also put some of you refresh and rendering commands into functions so they could be reused and organized. It's a bit messy; just put it together.